+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/hooks/useProgramStages.js
@@ -3,15 +3,15 @@ import { useMemo } from 'react';
 import log from 'loglevel';
 import { errorCreator } from 'capture-core-utils';
 import i18n from '@dhis2/d2-i18n';
-import type { apiProgramStage } from 'capture-core/metaDataStoreLoaders/programs/quickStoreOperations/types';
+import type { ProgramStages } from '../EnrollmentPageDefault.types';
 import { Program } from '../../../../../metaData';
 
-export const useProgramStages = (program: Program, programStages?: Array<apiProgramStage>) => useMemo(() => {
+export const useProgramStages = (program: Program, programStages?: ProgramStages) => useMemo(() => {
     const stages = [];
     if (program && programStages) {
         program.stages.forEach((item) => {
             const { id, name, icon, stageForm } = item;
-            const { hideDueDate, programStageDataElements, repeatable, enableUserAssignment } = programStages.find(p => p.id === id) || {};
+            const { hideDueDate, programStageDataElements, repeatable, enableUserAssignment, dataAccess } = programStages.find(p => p.id === id) || {};
             if (!programStageDataElements) {
                 log.error(errorCreator(i18n.t('Program stage not found'))(id));
             } else {
@@ -23,6 +23,7 @@ export const useProgramStages = (program: Program, programStages?: Array<apiProg
                     repeatable,
                     enableUserAssignment,
                     description: stageForm.description,
+                    dataAccess,
                     dataElements: programStageDataElements?.reduce((acc, currentStageData) => {
                         const { displayInReports, dataElement } = currentStageData;
                         if (displayInReports) {