+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/useBuildNewEventPayload.js
@@ -1,6 +1,6 @@
 // @flow
 import { useSelector } from 'react-redux';
-import { useTimeZoneConversion } from '@dhis2/app-runtime';
+import { useTimeZoneConversion, useConfig } from '@dhis2/app-runtime';
 import type { RenderFoundation } from '../../../metaData';
 import { getAddEventEnrollmentServerData } from './getConvertedAddEvent';
 import { convertDataEntryToClientValues } from '../../DataEntry/common/convertDataEntryToClientValues';
@@ -58,6 +58,7 @@ export const useBuildNewEventPayload = ({
     orgUnitName,
     formFoundation,
 }: Props) => {
+    const { serverVersion: { minor } } = useConfig();
     const dataEntryKey = `${dataEntryId}-${itemId}`;
     const formValues = useSelector(({ formsValues }) => formsValues[dataEntryKey]);
     const dataEntryValues = useSelector(({ dataEntriesFieldsValue }) => dataEntriesFieldsValue[dataEntryKey]);
@@ -144,6 +145,7 @@ export const useBuildNewEventPayload = ({
             orgUnitName,
             completed: saveType === addEventSaveTypes.COMPLETE,
             fromClientDate,
+            serverMinorVersion: minor,
         });
 
         const {